server/requirements.txt
praw
python-dotenv
requests
fastapi
uvicorn[standard]
pytrends
pydantic
pandas
sqlalchemy
apscheduler
httpx
tenacity
python-dateutil
cryptography